requirements.txt
blinker==1.9.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.2
fonttools==4.59.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.0
jsonpickle==4.1.1
keyboard==0.13.5
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mido==1.3.3
more-itertools==10.8.0
mss==10.1.0
music21==9.7.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pretty_midi==0.2.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-rtmidi==1.5.8
requests==2.32.5
setuptools==80.9.0
six==1.17.0
tkinterdnd2==0.4.3
toml==0.10.2
urllib3==2.5.0
uv==0.7.20
virtualenv==20.31.2
webcolors==24.11.1
Werkzeug==3.1.3
